25 Apr 2023, 11:24

+++ brushfire.cpp
@@ -118,8 +118,8 @@ void driveToPoint(vector<Pixel> points)
 
     for (int i = 0; i < points.size(); i++){
 
-        float x = 4000*(1-(points.at(i).x/128));
-        float y = 4000*((points.at(i).y/128));
+        double x = 4000*(1-(points.at(i).x/128));
+        double y = 4000*((points.at(i).y/128));
         int newDist = sqrt(x*x+y*y);
 
         printf("Pixel: (%i, %i) => (%f, %f), distFromPlayer; %i\n", points.at(i).x, points.at(i).y, x, y, dist);
